chore: Ajustar tipo de datos en DTOs

+++ src/main/java/com/musabeli/productor_ubicaciones/dtos/UbicacionRequestDto.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class UbicacionRequestDto {
     private String patente;
-    private Double latitud;
-    private Double longitud;
+    private BigDecimal latitud;
+    private BigDecimal longitud;
 }
